Use fd instead of /dev/std{out,err} (Fixes #2)
Writing to std{out,err} is done by writing to fd 1 or 2, not by writing to
/dev/std{out,err}. That's not the same.

This will fix the

	/dev/stderr: Permission denied

problem, when running this script using 'su' for example, because
/dev/stderr is a symlink to /proc/self/fd/2 on most Linux distributions
which is a symlink to /dev/{pts,tty} which isn't owned by the user who run
this script.

Again, /dev/std{in,out,err} is not the same like "stdin", "stdout" or
"stderr"!